Se hace refactor de cambios asociados a la estructura
