+++ lib/data/models/investment.dart
