+++ lib/data/providers/datasource_providers.dart
@@ -1,5 +1,5 @@
-import 'package:my_new_flutter_app/data/datasources/investment_local_datasource.dart';
-import 'package:my_new_flutter_app/data/datasources/investment_remote_datasource.dart';
+import 'package:my_new_flutter_app/domain/datasources/investment_local_datasource.dart';
+import 'package:my_new_flutter_app/domain/datasources/investment_remote_datasource.dart';
 import 'package:my_new_flutter_app/data/providers/dio_provider.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
