+++ lib/data/providers/get_investment_provider.dart
@@ -1,8 +1,8 @@
 import 'package:my_new_flutter_app/data/models/card_investment.dart';
-import 'package:my_new_flutter_app/data/repositories/investment_repository_contract.dart';
 import 'package:my_new_flutter_app/data/repositories/investment_repository.dart';
-import 'package:my_new_flutter_app/data/datasources/investment_local_datasource.dart';
-import 'package:my_new_flutter_app/data/datasources/investment_remote_datasource.dart';
+import 'package:my_new_flutter_app/domain/datasources/investment_control_datasource.dart';
+import 'package:my_new_flutter_app/domain/datasources/investment_local_datasource.dart';
+import 'package:my_new_flutter_app/domain/datasources/investment_remote_datasource.dart';
 import 'package:my_new_flutter_app/data/providers/dio_provider.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 part 'get_investment_provider.g.dart';
